fix size_t for strlen

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: yhasegaw <yhasegaw@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/07 01:37:30 by yhasegaw          #+#    #+#             */
-/*   Updated: 2020/10/28 01:42:12 by yhasegaw         ###   ########.fr       */
+/*   Updated: 2020/10/30 02:14:59 by yhasegaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,12 +18,12 @@ unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
 	unsigned int dst_len;
 
 	if (size == 0)
-		return (ft_strlen(src));
-	else if (size <= ft_strlen(dest))
-		return (size + ft_strlen(src));
+		return ((unsigned int)ft_strlen(src));
+	else if (size <= (unsigned int)ft_strlen(dest))
+		return (size + (unsigned int)ft_strlen(src));
 	else
 	{
-		dst_len = ft_strlen(dest);
+		dst_len = (unsigned int)ft_strlen(dest);
 		i = 0;
 		while (i < size - dst_len - 1 && src[i])
 		{
@@ -31,6 +31,6 @@ unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
 			i++;
 		}
 		dest[dst_len + i] = '\0';
-		return (dst_len + ft_strlen(src));
+		return (dst_len + (unsigned int)ft_strlen(src));
 	}
 }
